is_prime function working
got the base if the is_prime method working and returning true or false
based upon if the number is prime or not.  The method also only looks
for the next number it is divisible by beyond 2 (itself and 1 would make
it be 2) and then breaks the loop.  This should save on processing time